+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Constraint/Init.hs
@@ -118,7 +118,7 @@ makeAutoDecrDataCons dcts specenv dcs
     tycons      = L.nub $ mapMaybe idTyCon dcs
 
     go tycon
-      | S.member tycon specenv =  zipWith (makeSizedDataCons dcts) (tyConDataCons tycon) [0..]
+      | S.member tycon specenv =  zipWith (makeSizedDataCons specenv dcts) (tyConDataCons tycon) [0..]
     go _
       = []
 
@@ -131,15 +131,19 @@ idTyCon = fmap dataConTyCon . Ghc.isDataConId_maybe
 lenOf :: F.Symbol -> F.Expr
 lenOf x = F.mkEApp lenLocSymbol [F.EVar x]
 
-makeSizedDataCons :: [(Id, SpecType)] -> DataCon -> Integer -> (RSort, (Id, SpecType))
-makeSizedDataCons dcts x' n = (toRSort $ ty_res trep, (x, fromRTypeRep trep{ty_res = tres}))
+makeSizedDataCons :: S.HashSet TyCon -> [(Id, SpecType)] -> DataCon -> Integer -> (RSort, (Id, SpecType))
+makeSizedDataCons specenv dcts x' n = (toRSort $ ty_res trep, (x, fromRTypeRep trep{ty_res = tres}))
     where
       x      = dataConWorkId x'
       st     = fromMaybe (impossible Nothing "makeSizedDataCons: this should never happen") $ L.lookup x dcts
       trep   = toRTypeRep st
       tres   = ty_res trep `strengthen` MkUReft (F.Reft (F.vv_, F.PAtom F.Eq (lenOf F.vv_) computelen)) mempty
 
-      recarguments = filter (\(t,_) -> toRSort t == toRSort tres) (zip (ty_args trep) (ty_binds trep))
+      -- An argument of an autosized type contributes its own size, so the size
+      -- of a value grows along the edges between mutually recursive types too.
+      recarguments = filter (autosized . fst) (zip (ty_args trep) (ty_binds trep))
+      autosized (RApp c _ _ _) = S.member (rtc_tc c) specenv
+      autosized _              = False
       computelen   = foldr (F.EBin F.Plus) (F.ECon $ F.I n) (lenOf .  snd <$> recarguments)
 
 mergeDataConTypes ::  F.TCEmb TyCon -> [(Var, SpecType)] -> [(Var, SpecType)] -> [(Var, SpecType)]